requirements.txt
# Slack Integration
slack-sdk==3.27.1
slack-bolt==1.18.1

# Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.25  # Optional: ORM if you prefer

# Environment & Configuration
python-dotenv==1.0.1

# Async Support
aiohttp==3.9.3
asyncio==3.4.3

# Utilities
requests==2.31.0
urllib3==2.2.0

# Logging & Monitoring
python-json-logger==2.0.7

# CLI
click==8.1.7  # For building CLI scripts

# Development (optional)
pytest==8.0.0
pytest-asyncio==0.23.5
black==24.2.0
flake8==7.0.0
